+++ Packages/com.awavr.hierarchyplus/CHANGELOG.md
@@ -1,3 +1,9 @@
-(v1.3.0)
+(v1.3.2)
+--------
+- [FIX]: Layer and Tag labels not showing when there's a gameobject with a lot of icons, and having a normal sized Hierarchy panel.
+- [FEAT]: Left align Layer and Tag labels.
+- [FEAT]: Add tooltips to certain menu icons to clarify what they do.
+
+(v1.3.1)
 --------
 - [FIX]: Icons sometimes not loading when loading the editor.